requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.8
django-bootstrap3==9.0.0
django-countries==4.6.1
django-debug-toolbar==1.8
django-extensions==1.9.1
django-geoposition==0.3.0
django-jquery==3.1.0
django-markdown==0.8.4
django-static-jquery==2.1.4
django-toolbelt==0.0.1
gunicorn==19.7.1
Markdown==2.6.8
pkg-resources==0.0.0
psycopg2==2.7.3
pytz==2016.10
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
whitenoise==3.3.1
